refactor : entity에 dto의존성 제거
