+++ hashtagmap-core/src/main/java/com/songpapeople/hashtagmap/kakao/schedule/repository/PeriodHistoryRepository.java
@@ -1,5 +1,6 @@
-package com.songpapeople.hashtagmap.kakao.schedule;
+package com.songpapeople.hashtagmap.kakao.schedule.repository;
 
+import com.songpapeople.hashtagmap.kakao.schedule.model.PeriodHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PeriodHistoryRepository extends JpaRepository<PeriodHistory, Long> {
